修复退出登录时，删除待办事项是仍会删除云数据的bug

+++ app/src/main/java/com/example/lulin/todolist/fragment/TodoFragment.java
@@ -112,17 +112,20 @@ public class TodoFragment extends Fragment {
 //                                    }
 //                                });
 
-                                ToDoUtils.deleteNetTodos(getContext(), todos, new ToDoUtils.DeleteTaskListener() {
-                                    @Override
-                                    public void onSuccess() {
+                                if (User.getCurrentUser(User.class) != null){
+                                    ToDoUtils.deleteNetTodos(getContext(), todos, new ToDoUtils.DeleteTaskListener() {
+                                        @Override
+                                        public void onSuccess() {
 
-                                    }
+                                        }
 
-                                    @Override
-                                    public void onError(int errorCord, String msg) {
-                                        ToastUtils.showShort(getContext(),msg);
-                                    }
-                                });
+                                        @Override
+                                        public void onError(int errorCord, String msg) {
+                                            ToastUtils.showShort(getContext(),msg);
+                                        }
+                                    });
+                                }
+
                                 todoRecyclerViewAdapter.removeItem(position);
 
                             }
